+++ angular_compiler/lib/angular_compiler.dart
@@ -50,5 +50,4 @@ export 'src/analyzer.dart'
 export 'src/asset.dart' show NgAssetReader;
 export 'src/emitter/injector.dart' show InjectorEmitter;
 export 'src/emitter/reflector.dart' show ReflectableEmitter, SplitDartEmitter;
-export 'src/flags.dart' show CompilerFlags, Profile;
 export 'src/outliner.dart' show TemplateOutliner;