+++ angular_compiler/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.4.0-alpha+6
+
 ### New features
 
 *   Added an internal `cli.dart` library. See `lib/cli.dart` for details.